pyproject.toml
[project]
name = "probe"
version = "0.1.0"
description = "Read-only log troubleshooting MCP service"
readme = "ROADMAP.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "mcp>=1.24.0",
  "sse-starlette>=2.1.3",
  "PyYAML>=6.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
